Freshen up dependnecies and simplify build system
This commit makes a few changes to freshen up and simplify the code
repository. First, it upgrades the docker-compose configuration to
use node 8 instead of node 6. Next, it removes the dependency on
Bower, which is now outdated, and it also removes the need for Gulp
by using straight NPM scripts.