Update packages and remove Docker Compose config
This commit makes a bit of cleanup to the repository. First it removes
the docker-compose.yml file and adds a single Dockerfile instead.
Second, it updates the major versions of outdated NPM packages to
improve security and performance of the site. Finally, the README is
updated to show the non-Docker build commands.